requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
python-dotenv
python-jose
passlib[bcrypt]
sqlalchemy
jinja2
